Flatten the country selection modal styles
